requirements.txt
fastapi
uvicorn[standard]

accelerate==1.13.0
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
async-timeout==5.0.1
attrs==26.1.0
beautifulsoup4==4.14.3
bitsandbytes==0.47.0
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==46.0.7
dataclasses-json==0.6.7
defusedxml==0.7.1
Deprecated==1.3.1
distro==1.9.0
ecdsa==0.19.2
filelock==3.29.0
filetype==1.2.0
frozenlist==1.8.0
fsspec==2026.3.0

google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.73.1
google-generativeai==0.8.6
googleapis-common-protos==1.74.0

greenlet==3.4.0
grpcio==1.80.0
grpcio-status==1.71.2

huggingface_hub==0.36.2
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1

langchain-core==0.3.84
langchain-huggingface==0.3.1
langsmith==0.7.34

llama-index==0.14.8
llama-index-core==0.14.21
llama-index-embeddings-huggingface==0.6.1
llama-index-embeddings-openai==0.5.2
llama-index-llms-openai==0.6.26

marshmallow==3.26.2
multidict==6.7.1
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.4

numpy==2.2.6
openai==2.32.0
orjson==3.11.8
packaging==25.0
pandas==2.3.3
pillow==10.4.0
protobuf==5.29.6
psutil==7.2.2

pydantic
pydantic-settings

python-dateutil==2.9.0.post0
<<<<<<< HEAD
python-dotenv
python-multipart

regex==2026.4.4
requests==2.33.1

=======
python-jose==3.5.0
pytz==2026.1.post1
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rsa==4.9.1
>>>>>>> 8404d7a06383f56cdf512e37478e94f83af1ce05
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sentence-transformers==5.4.1

SQLAlchemy==2.0.49
sympy==1.14.0
tenacity==9.1.4

tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
transformers==4.56.1

tqdm==4.67.3
typing-inspect==0.9.0
tzdata==2026.1

urllib3==2.6.3
wrapt==2.1.2
<<<<<<< HEAD
yarl==1.23.0
=======
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
>>>>>>> 8404d7a06383f56cdf512e37478e94f83af1ce05
